Implement a basic blockEntity system and implement Furnace inventar (without functions) and change something

+++ src/main/java/org/jukeboxmc/network/packet/ContainerClosePacket.java
@@ -12,7 +12,7 @@ import org.jukeboxmc.network.Protocol;
 @EqualsAndHashCode ( callSuper = true )
 public class ContainerClosePacket extends Packet {
 
-    private byte windowId;
+    private int windowId;
     private boolean isServerInitiated;
 
     @Override
